Update- work on onboarding/utils
