+++ package.json
@@ -3,15 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "lodash-es": "4.17.11",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "axios": "^0.19.2",
+    "lodash-es": "4.17.11",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-scripts": "3.3.1",
     "react-spring": "^8.0.27",
-    "react-use-gesture": "^7.0.1"
+    "react-use-gesture": "^7.0.1",
+    "redux": "^4.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
