+++ src/App.css
@@ -1,0 +1,231 @@
+#root,
+html,
+body {
+  max-width: 100vw;
+  margin: 0 auto;
+}
+
+.App {
+  width: 100vw;
+  height: auto;
+  display: flex;
+  flex-flow: column nowrap;
+  font-family: "Poiret One", cursive;
+  background-color: #181818;
+}
+
+.app-nav {
+  background-color: #282c34;
+  position: fixed;
+  width: 100vw;
+  height: 100px;
+  display: flex;
+  flex-flow: row nowrap;
+  justify-content: flex-end;
+  font-size: calc(10px + 2vmin);
+  margin-bottom: 40px;
+  z-index: 1;
+}
+
+.app-nav a {
+  text-decoration: none;
+  color: #ffffff;
+  font-size: 28pt;
+  font-weight: bold;
+  flex-grow: 80;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 80%;
+  margin-right: 2rem;
+  border-radius: 2px;
+}
+
+.app-nav a:hover {
+  background-color: #ffffff;
+  color: #000000;
+}
+
+.links {
+  display: flex;
+  width: 20%;
+  height: 100%;
+  justify-content: space-between;
+  align-items: center;
+  margin-right: 0.5rem;
+}
+
+.login-nav {
+  margin-right: 60%;
+  width: 13%;
+}
+
+.login-nav button {
+  display: block;
+  margin-top: 0.3rem;
+  margin-left: 3rem;
+}
+
+.hero {
+  background-image: url("./Assets/friends-splash.png");
+  background-size: 100% 100%;
+  background-repeat: no-repeat;
+  height: 850px;
+  width: 100%;
+  margin-top: 100px;
+  text-align: center;
+  color: #ffffff;
+  font-size: 31px;
+}
+
+.hero h1 {
+  font-size: 80pt;
+  font-weight: 700;
+  margin-top: 1.5rem;
+  color: #b62eaf;
+}
+
+.friends-list-wrapper {
+  position: relative;
+  margin-top: 100px;
+  display: flex;
+  width: 100vw;
+  height: auto;
+  flex-flow: column wrap;
+  align-items: center;
+  margin: 0 auto;
+  z-index: 0;
+}
+
+.friends-list {
+  margin-top: 100px;
+  padding-left: 0;
+  width: auto;
+}
+
+.friend-card {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-items: center;
+  width: 35rem;
+  height: 25rem;
+  border: 15px solid rgb(75, 16, 16);
+  border-radius: 10px;
+  margin: 0 auto;
+  background-color: #ffffff;
+  opacity: 0.75;
+  margin-top: 2rem;
+  padding-bottom: 0.5rem;
+}
+
+.friend-card h1 {
+  font-weight: bold;
+  font-size: 51px;
+}
+
+.update-form {
+  display: flex;
+  flex-flow: column nowrap;
+  justify-content: center;
+  align-items: center;
+  background-color: rgb(82, 35, 87);
+  height: 100vh;
+  width: 80vw;
+  color: #ffffff;
+  margin: 0 auto;
+  box-shadow: purple 10px 10px 30px;
+}
+
+.update-input {
+  margin-right: 1.5rem;
+  margin-top: 1rem;
+}
+
+.spinner-center {
+  color: #ffffff;
+  width: auto;
+  margin: 0 auto;
+}
+
+.spinner {
+  display: inline-block;
+  margin-top: 200px;
+  border-radius: 500px;
+  margin-left: 20rem;
+}
+
+.md-button {
+  margin: 0.5rem;
+  padding: 0.5rem;
+  width: 150px;
+  font-size: 20px;
+}
+
+@media only screen and (max-width: 750px) {
+  .App {
+    width: auto;
+  }
+
+  .links {
+    display: flex;
+    width: 40%;
+    height: 100%;
+    justify-items: center;
+    align-items: center;
+    margin-right: 0.5rem;
+    margin-left: 4rem;
+  }
+
+  .app-nav {
+    background-color: #282c34;
+    position: fixed;
+    width: 100vw;
+    height: 150px;
+    display: flex;
+    flex-flow: columnn nowrap;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    margin-bottom: 40px;
+    z-index: 1;
+  }
+
+  .app-nav a {
+    text-decoration: none;
+    color: #ffffff;
+    font-size: 28pt;
+    font-weight: bold;
+    flex-grow: 80;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 80%;
+    margin: 0.5rem;
+    border-radius: 2px;
+  }
+
+  .login-nav {
+    width: 10rem;
+    margin: 2rem;
+  }
+
+  .hero {
+    background-image: url("./Assets/road-friends.png");
+    background-size: 100% 100%;
+    background-repeat: no-repeat;
+    height: 850px;
+    width: 100%;
+    margin-top: 150px;
+    text-align: center;
+    color: #ffffff;
+    font-size: 31px;
+  }
+
+  .hero h1 {
+    font-size: 80pt;
+    font-weight: 700;
+    margin-top: 1.5rem;
+    color: #b62eaf;
+  }
+}
